차량 상세 화면 api 연동 및 Firebase Realtime Database 연동

+++ Trever/Features/Home/Models/CarListItem.swift
@@ -2,7 +2,7 @@ import Foundation
 
 struct CarListItem: Identifiable, Hashable {
     let id: UUID
-    let backendId: Int64?
+    let backendId: Int?
     let title: String
     let subTitle: String?
     let year: Int
